+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 Release 1.0.1
+* exception is thrown when insertion of byte code into a method does not find insertion point.
+* can detect method invocation within a method.
+
+Release 1.0.1
 
 * fix bug in afterburner sample